move ExperimentResultsWriter to seperate file
